+++ src/main/java/com/example/hrm_be/services/impl/PurchaseServiceImpl.java
@@ -5,9 +5,6 @@ import com.example.hrm_be.components.PurchaseMapper;
 import com.example.hrm_be.configs.exceptions.HrmCommonException;
 import com.example.hrm_be.models.dtos.Purchase;
 import com.example.hrm_be.models.entities.PurchaseEntity;
-import com.example.hrm_be.models.entities.SupplierEntity;
-import com.example.hrm_be.models.requests.purchase.PurchaseCreateRequest;
-import com.example.hrm_be.models.requests.purchase.PurchaseUpdateRequest;
 import com.example.hrm_be.repositories.PurchaseRepository;
 import com.example.hrm_be.services.PurchaseService;
 import io.micrometer.common.util.StringUtils;
@@ -45,43 +42,33 @@ public class PurchaseServiceImpl implements PurchaseService {
   }
 
   @Override
-  public Purchase create(PurchaseCreateRequest Purchase) {
-    if (Purchase == null) {
-      throw new HrmCommonException(HrmConstant.ERROR.BRANCH.EXIST);
-    }
-
-    SupplierEntity supplier;
-    if (Purchase.getSupplierId() != null) {
-      supplier = entityManager.getReference(SupplierEntity.class, Purchase.getSupplierId());
-      if (supplier == null) {
-        throw new HrmCommonException("Supplier not found with id: " + Purchase.getSupplierId());
-      }
-    } else {
-      supplier = null;
+  public Purchase create(Purchase purchase) {
+    if (purchase == null) {
+      throw new HrmCommonException(HrmConstant.ERROR.PURCHASE.EXIST);
     }
 
     // Convert DTO to entity, save it, and convert back to DTO
-    return Optional.ofNullable(Purchase)
-        .map(e -> purchaseMapper.toEntity(e, supplier))
+    return Optional.ofNullable(purchase)
+        .map(purchaseMapper::toEntity)
         .map(e -> purchaseRepository.save(e))
         .map(e -> purchaseMapper.toDTO(e))
         .orElse(null);
   }
 
   @Override
-  public Purchase update(PurchaseUpdateRequest Purchase) {
-    PurchaseEntity oldPurchaseEntity = purchaseRepository.findById(Purchase.getId()).orElse(null);
+  public Purchase update(Purchase purchase) {
+    PurchaseEntity oldPurchaseEntity = purchaseRepository.findById(purchase.getId()).orElse(null);
     if (oldPurchaseEntity == null) {
-      throw new HrmCommonException(HrmConstant.ERROR.BRANCH.NOT_EXIST);
+      throw new HrmCommonException(HrmConstant.ERROR.PURCHASE.NOT_EXIST);
     }
 
     return Optional.ofNullable(oldPurchaseEntity)
         .map(
             op ->
                 op.toBuilder()
-                    .amount(Purchase.getAmount())
-                    .remainDebt(Purchase.getRemainDebt())
-                    .purchaseDate(Purchase.getPurchaseDate())
+                    .amount(purchase.getAmount())
+                    .remainDebt(purchase.getRemainDebt())
+                    .purchaseDate(purchase.getPurchaseDate())
                     .build())
         .map(purchaseRepository::save)
         .map(purchaseMapper::toDTO)
@@ -92,6 +79,11 @@ public class PurchaseServiceImpl implements PurchaseService {
   public void delete(Long id) {
     if (StringUtils.isBlank(id.toString())) {
       return;
+    }
+
+    PurchaseEntity oldPurchaseEntity = purchaseRepository.findById(id).orElse(null);
+    if (oldPurchaseEntity == null) {
+      throw new HrmCommonException(HrmConstant.ERROR.PURCHASE.NOT_EXIST);
     }
 
     purchaseRepository.deleteById(id);
